+++ src/main/java/com/neonatal/backend/repositories/RulesRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
  * Repository interface for the Rules Design on the front-end.  Allows DAO to the database table "rules"
  */
 @Repository
-public interface RulesRepository extends JpaRepository<Rule, Integer> {
+public interface RulesRepository extends JpaRepository<Rule, Long> {
 
     /*
     @Query(value = "SELECT * FROM rules where id = :id", nativeQuery = true)
